연관관계 다시 수정, PreRemove 추가, 테스트케이스 수정

+++ src/main/java/com/youngchayoungcha/tastynote/web/dto/PostCreateDTO.java
@@ -18,7 +18,7 @@ public class PostCreateDTO {
 
     private String content;
 
-    private short score;
+    private float score;
 
     private boolean isPublic;
