+++ src/main/java/com/youngchayoungcha/tastynote/web/dto/PostResponseDTO.java
@@ -17,7 +17,7 @@ public class PostResponseDTO {
 
     private String content;
 
-    private short score;
+    private Float score;
 
     private boolean isPublic;
